+++ CHANGES.md
@@ -1,6 +1,25 @@
 videojs-record changelog
 ========================
 
+3.6.0 - unreleased
+------------------
+
+- Add support for keyboard hotkeys (requires video.js 7.5.0 or newer) (#339)
+- Trigger error when `getUserMedia` or `getDisplayMedia` support is missing (in
+  Chrome iOS for example) (#353)
+- Improve pause/resume accuracy with monotonic clock
+- Make sure recorded data is garbage collected
+- Bump required version for videojs-wavesurfer to 2.8.0 for wavesurfer.js fixes
+- Examples: add `playsinline` attribute to `video` element for Safari (#353)
+
+
+3.5.1 - 2019/03/06
+------------------
+
+- Fix RecordRTC import (#345)
+- Update Safari `AudioContext` workaround in examples (#335)
+
+
 3.5.0 - 2019/02/28
 ------------------